@@ -1,10 +1,6 @@
 <template>
   <div id="app">
-    <div id="nav">
-      <router-link to="/">Home</router-link> |
-      <router-link to="/about">About</router-link> |
-      <router-link to="/calendar">Calendar</router-link>
-    </div>
+    <Navbar></Navbar>
     <router-view/>
   </div>
 </template>
@@ -30,3 +26,9 @@
   color: #42b983;
 }
 </style>
+<script>
+  import Navbar from "./components/Navbar";
+  export default {
+    components: {Navbar}
+  }
+</script>
