@@ -1,105 +1,111 @@
 'use strict'
 
-import { app, protocol, BrowserWindow, dialog } from 'electron'
+import {app, protocol, BrowserWindow, dialog} from 'electron'
 import {
-  createProtocol,
-  /* installVueDevtools */
+    createProtocol,
+    /* installVueDevtools */
 } from 'vue-cli-plugin-electron-builder/lib'
+
 const isDevelopment = process.env.NODE_ENV !== 'production'
-const { ipcMain } = require('electron');
+const {ipcMain} = require('electron');
+const path = require('path');
 
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
 let win
 
 // Scheme must be registered before the app is ready
-protocol.registerSchemesAsPrivileged([{scheme: 'app', privileges: { secure: true, standard: true } }])
+protocol.registerSchemesAsPrivileged([{scheme: 'app', privileges: {secure: true, standard: true}}])
 
-function createWindow () {
-  // Create the browser window.
-  win = new BrowserWindow({ width: 800, height: 600, webPreferences: {
-    nodeIntegration: true
-  } })
+function createWindow() {
+    // Create the browser window.
+    win = new BrowserWindow(
+        {
+            width: 1024, height: 800, minWidth: 1024, webPreferences: {
+                nodeIntegration: process.env.ELECTRON_NODE_INTEGRATION,
+                preload: path.join(__dirname, 'preload.js')
+            }
+        })
 
-  if (process.env.WEBPACK_DEV_SERVER_URL) {
-    // Load the url of the dev server if in development mode
-    win.loadURL(process.env.WEBPACK_DEV_SERVER_URL)
-    if (!process.env.IS_TEST) win.webContents.openDevTools()
-  } else {
-    createProtocol('app')
-    // Load the index.html when not in development
-    win.loadURL('app://./index.html')
-  }
+    if (process.env.WEBPACK_DEV_SERVER_URL) {
+        // Load the url of the dev server if in development mode
+        win.loadURL(process.env.WEBPACK_DEV_SERVER_URL)
+        if (!process.env.IS_TEST) win.webContents.openDevTools()
+    } else {
+        createProtocol('app')
+        // Load the index.html when not in development
+        win.loadURL('app://./index.html')
+    }
 
-  win.on('closed', () => {
-    win = null
-  })
+    win.on('closed', () => {
+        win = null
+    })
 }
 
 // Quit when all windows are closed.
 app.on('window-all-closed', () => {
-  // On macOS it is common for applications and their menu bar
-  // to stay active until the user quits explicitly with Cmd + Q
-  if (process.platform !== 'darwin') {
-    app.quit()
-  }
+    // On macOS it is common for applications and their menu bar
+    // to stay active until the user quits explicitly with Cmd + Q
+    if (process.platform !== 'darwin') {
+        app.quit()
+    }
 })
 
 app.on('activate', () => {
-  // On macOS it's common to re-create a window in the app when the
-  // dock icon is clicked and there are no other windows open.
-  if (win === null) {
-    createWindow()
-  }
+    // On macOS it's common to re-create a window in the app when the
+    // dock icon is clicked and there are no other windows open.
+    if (win === null) {
+        createWindow()
+    }
 })
 
 // This method will be called when Electron has finished
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
 app.on('ready', async () => {
-  if (isDevelopment && !process.env.IS_TEST) {
-    // Install Vue Devtools
-    // Devtools extensions are broken in Electron 6.0.0 and greater
-    // See https://github.com/nklayman/vue-cli-plugin-electron-builder/issues/378 for more info
-    // Electron will not launch with Devtools extensions installed on Windows 10 with dark mode
-    // If you are not using Windows 10 dark mode, you may uncomment these lines
-    // In addition, if the linked issue is closed, you can upgrade electron and uncomment these lines
-    // try {
-    //   await installVueDevtools()
-    // } catch (e) {
-    //   console.error('Vue Devtools failed to install:', e.toString())
-    // }
+    if (isDevelopment && !process.env.IS_TEST) {
+        // Install Vue Devtools
+        // Devtools extensions are broken in Electron 6.0.0 and greater
+        // See https://github.com/nklayman/vue-cli-plugin-electron-builder/issues/378 for more info
+        // Electron will not launch with Devtools extensions installed on Windows 10 with dark mode
+        // If you are not using Windows 10 dark mode, you may uncomment these lines
+        // In addition, if the linked issue is closed, you can upgrade electron and uncomment these lines
+        // try {
+        //   await installVueDevtools()
+        // } catch (e) {
+        //   console.error('Vue Devtools failed to install:', e.toString())
+        // }
 
-  }
-  createWindow()
+    }
+    createWindow()
 })
 
 // Exit cleanly on request from parent process in development mode.
 if (isDevelopment) {
-  if (process.platform === 'win32') {
-    process.on('message', data => {
-      if (data === 'graceful-exit') {
-        app.quit()
-      }
-    })
-  } else {
-    process.on('SIGTERM', () => {
-      app.quit()
-    })
-  }
+    if (process.platform === 'win32') {
+        process.on('message', data => {
+            if (data === 'graceful-exit') {
+                app.quit()
+            }
+        })
+    } else {
+        process.on('SIGTERM', () => {
+            app.quit()
+        })
+    }
 }
 
-ipcMain.on('show-open-dialog', (event, year, month, day)=> {
-  console.log(`Open File Dialog for ${year} ${month} ${day}`);
-  let filePaths = dialog.showOpenDialog({
-    title: "Choose a video or image",
-    filters: [
-      { name: 'All media files', extensions: ['mp4', 'mov', 'avi', 'mpg', 'mpeg', 'jpg', 'jpeg', 'gif', 'png']},
-      { name: 'Videos', extensions: ['mp4', 'mov', 'avi', 'mpg', 'mpeg']},
-      { name: 'Images', extensions: ['jpg', 'jpeg', 'gif', 'png']}
-    ],
-    properties: ['openFile']
-  });
+ipcMain.on('show-open-dialog', (event, year, month, day) => {
+    console.log(`Open File Dialog for ${year} ${month} ${day}`);
+    let filePaths = dialog.showOpenDialog({
+        title: "Choose a video or image",
+        filters: [
+            {name: 'All media files', extensions: ['mp4', 'mov', 'avi', 'mpg', 'mpeg', 'jpg', 'jpeg', 'gif', 'png']},
+            {name: 'Videos', extensions: ['mp4', 'mov', 'avi', 'mpg', 'mpeg']},
+            {name: 'Images', extensions: ['jpg', 'jpeg', 'gif', 'png']}
+        ],
+        properties: ['openFile']
+    });
     let filePath = filePaths[0];
     console.log(`File path ${filePath}`);
 });
