@@ -1,16 +1,10 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import Home from './views/Home.vue'
 
 Vue.use(Router)
 
 export default new Router({
   routes: [
-    {
-      path: '/',
-      name: 'home',
-      component: Home
-    },
     {
       path: '/about',
       name: 'about',
@@ -22,10 +16,10 @@ export default new Router({
       }
     },
     {
-      path: '/calendar',
+      path: '/',
       name: 'calendar',
       component: function () {
-        return import(/* webpackChunkName: "about" */ './views/Calendar.vue')
+        return import(/* webpackChunkName: "calendar" */ './views/Calendar.vue')
       }
     }
   ]
