@@ -1,13 +1,21 @@
 <template>
     <div>
-        <CalendarMonth v-bind:month="3" v-bind:year="2020"></CalendarMonth>
+        <MonthSelector v-bind:month=month v-bind:year=year></MonthSelector>
+        <CalendarMonth v-bind:month=month v-bind:year=year></CalendarMonth>
     </div>
 </template>
 <script>
     import CalendarDay from "../components/calendar/CalendarDay";
     import CalendarWeek from "../components/calendar/CalendarWeek";
     import CalendarMonth from "../components/calendar/CalendarMonth";
+    import MonthSelector from "../components/calendar/MonthSelector";
     export default {
-        components: {CalendarMonth, CalendarDay, CalendarWeek}
+        components: {MonthSelector, CalendarMonth, CalendarDay, CalendarWeek},
+        data: function () {
+            return {
+                month: 3,
+                year: 2020
+            }
+        }
     }
 </script>
